ship notebook extension json within the package
This resolves the issue of needing a separated environment,
because the notebook server is started with the configs
from the package.

Also cleaned up the structure a bit.
