+++ pyemma_tutorials/__init__.py
@@ -2,8 +2,8 @@
 Jupyter notebook launcher for PyEMMA's tutorials series.
 """
 
-from .__main__ import notebook_location, main
-
 from ._version import get_versions
 __version__ = get_versions()['version']
 del get_versions
+
+from .util import notebook_location, configs_location
